clock speed

+++ src/opencl/kernel/matmul1.cl
@@ -1,15 +1,15 @@
 // TODO implement type agnostic kernel
-__kernel void matmul1(__global const uint *lhs,
+__kernel void matmul1(__global const float *lhs,
   __global const size_t *lhs_dlen,
-  __global const uint *rhs,
+  __global const float *rhs,
   __global const size_t *rhs_dlen,
-  __global uint *res,
+  __global float *res,
   const size_t K /* col size of lhs, row size of rhs */ ) {
 
   const int row_index = get_global_id(0);
   const int col_index = get_global_id(1);
 
-  uint acc = 0;
+  float acc = 0.0f;
 
   for (size_t k = 0; k < K; k++) {
     acc += lhs[row_index + k * lhs_dlen[0]] * rhs[col_index * K + k];
